Move poll models into separate file
